+++ R/simList-accessors.R
@@ -1,7 +1,6 @@
 if (getRversion() >= "3.1.0") {
-  utils::globalVariables(c(".SD", "eventTime", "savetime", "exts",
-                           "eventType", "unit", "diffTime", "clockTime",
-                           "minEventTime", "maxEventTime", "eventNumber"))
+  utils::globalVariables(c(".SD", "clockTime", "diffTime", "eventNumber", "eventTime", "eventType",
+                           "exts", "minEventTime", "maxEventTime", "savetime", "unit"))
 }
 
 ### `show` generic is already defined in the methods package
@@ -122,7 +121,7 @@ setMethod(
 #' @aliases simList-accessors-envir
 #' @author Alex Chubaty
 #' @export
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @include simList-class.R
 #' @rdname simList-accessors-envir
 #'
@@ -175,7 +174,7 @@ setReplaceMethod("envir",
 #'
 #' @return Returns or sets a list of objects in the \code{simList} environment.
 #'
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @seealso \code{\link{SpaDES.core-package}}, specifically the section 1.2.1 on Simulation Parameters.
 #'
 #' @export
@@ -254,7 +253,7 @@ setReplaceMethod(
 #' @param hidden Logical. If TRUE, show the default core modules.
 #' @return Returns or sets the value of the slot from the \code{simList} object.
 #'
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @seealso \code{\link{SpaDES.core-package}}, specifically the section 1.2.7 on Modules and dependencies.
 #'
 #' @export
@@ -426,7 +425,7 @@ setMethod(
 #' part for speed (e.g., 2-4x faster) is specifying the moduleName.
 #' Specifying the parameter name is <5% faster.
 #'
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @seealso \code{\link{SpaDES.core-package}}, specifically the section 1.2.1 on Simulation parameters.
 #'
 #' @export
@@ -519,7 +518,7 @@ P <- function(sim, module, param) {
 #' @export
 #' @rdname globals
 #'
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @seealso \code{\link{SpaDES.core-package}}, specifically the section 1.2.1 on Simulation Parameters.
 #'
 setGeneric("globals", function(sim) {
@@ -587,6 +586,9 @@ setReplaceMethod("G",
 })
 
 ################################################################################
+#' @details
+#' \code{parameters} will extract only the metadata with the metadata defaults,
+#' NOT the current values that may be overwritten by a user. See examples.
 #' @inheritParams params
 #' @param asDF Logical. For \code{parameters}, if TRUE, this will produce a single
 #'                 data.frame of all model parameters. If FALSE, then it will return
@@ -601,8 +603,26 @@ setReplaceMethod("G",
 #' paths <- list(modulePath = system.file("sampleModules", package = "SpaDES.core"))
 #' mySim <- simInit(modules = modules, paths = paths,
 #'                  params = list(.globals = list(stackName = "landscape")))
-#' parameters(mySim)
 #'
+#' # update some parameters using assignment -- currently only params will work
+#' params(mySim)$randomLandscapes$nx <- 200
+#' params(mySim)$randomLandscapes$ny <- 200
+#'
+#' parameters(mySim) # Does not contain these user overridden values
+#'
+#' # These next 2 are same here because they are not within a module
+#' P(mySim)          # Does contain the user overridden values
+#' params(mySim)     # Does contain the user overridden values
+#'
+#' # NOTE -- deleting a parameter will affect params and P, not parameters
+#' params(mySim)$randomLandscapes$nx <- NULL
+#' params(mySim)$randomLandscapes$ny <- NULL
+#'
+#' parameters(mySim) # Shows nx and ny
+#'
+#' # These next 2 are same here because they are not within a module
+#' P(mySim)          # nx and ny are Gone
+#' params(mySim)     # nx and ny are Gone
 setGeneric("parameters", function(sim, asDF = FALSE) {
   standardGeneric("parameters")
 })
@@ -637,7 +657,7 @@ setMethod("parameters",
 #' @export
 #' @include simList-class.R
 #' @rdname checkpoint
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #'
 setGeneric("checkpointFile", function(sim) {
   standardGeneric("checkpointFile")
@@ -724,7 +744,7 @@ setReplaceMethod("checkpointInterval",
 #' @inheritParams params
 #' @include simList-class.R
 #' @export
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @rdname progress
 #'
 #' @examples
@@ -907,7 +927,7 @@ setReplaceMethod("progressType",
 #' @return Returns or sets the value(s) of the \code{input} or \code{output} slots
 #' in the \code{simList} object.
 #'
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @seealso \code{\link{SpaDES.core-package}}, specifically the section 1.2.2 on loading and saving.
 #'
 #' @include simList-class.R
@@ -1477,7 +1497,7 @@ setReplaceMethod(
 #'
 #' @return Returns or sets the value of the slot from the \code{simList} object.
 #'
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @seealso \code{\link{SpaDES.core-package}}, specifically the section 1.2.4 on Simulation Paths.
 #'
 #' @include simList-class.R
@@ -1809,7 +1829,7 @@ setMethod("dataPath",
 #'
 #' @return Returns or sets the value of the slot from the \code{simList} object.
 #'
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @seealso \code{\link{SpaDES.core-package}}, specifically the section 1.2.5 on Simulation times.
 #'
 #' @export
@@ -2223,7 +2243,7 @@ setMethod(
 #'
 #' @aliases simList-accessors-events
 #' @export
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @importFrom data.table := copy data.table
 #' @importFrom stats setNames
 #' @include simList-class.R
@@ -2294,7 +2314,7 @@ setReplaceMethod(
      }
 
      if (NROW(value)) {
-       sim@events <- lapply(seq_along(1:NROW(value)), function (x) as.list(value[x,]))
+       sim@events <- lapply(seq_along(1:NROW(value)), function(x) as.list(value[x,]))
      } else {
        sim@events <- list()
      }
@@ -2317,7 +2337,7 @@ setMethod(
   "conditionalEvents",
   signature = c("simList", "character"),
   definition = function(sim, unit) {
-    if (length(unit) != 1) stop ("unit must be length 1")
+    if (length(unit) != 1) stop("unit must be length 1")
     if (exists("._conditionalEvents", envir = sim, inherits = FALSE)) {
       conds <- sim$._conditionalEvents
       conds <- lapply(conds, function(x) {
@@ -2344,7 +2364,7 @@ setMethod(
     } else {
       return(NULL)
     }
-  })
+})
 
 #' @export
 #' @rdname simList-accessors-events
@@ -2354,7 +2374,7 @@ setMethod("conditionalEvents",
           definition = function(sim, unit) {
             res <- conditionalEvents(sim, sim@simtimes[["timeunit"]])
             return(res)
-          })
+})
 
 ################################################################################
 #' @inheritParams events
@@ -2444,7 +2464,7 @@ setGeneric("completed", function(sim, unit, times = TRUE) {
 #' @rdname simList-accessors-events
 #' @export
 #' @aliases simList-accessors-events
-#' @importFrom data.table rbindlist setkeyv :=
+#' @importFrom data.table rbindlist set setkeyv :=
 setMethod(
   "completed",
   signature = c("simList", "character"),
@@ -2457,16 +2477,16 @@ setMethod(
       obj[, eventNumber := as.numeric(eventNumber)]
       setkeyv(obj, "eventNumber")
       if (!isTRUE(times)) {
-        set(obj, , "._clockTime", NULL)
+        set(obj, NULL, "._clockTime", NULL)
       }
       if (is.na(pmatch("second", unit)) & (length(sim@completed))) {
         # note the above line captures empty eventTime, whereas `is.na` does not
         if (any(!is.na(obj$eventTime))) {
           if (!is.null(obj$eventTime)) {
             if (!is.null(obj$._clockTime))
-              obj[, `:=`(eventTime=convertTimeunit(eventTime, unit, sim@.xData),
-                         clockTime=obj$._clockTime,
-                         ._clockTime=NULL)]
+              obj[, `:=`(eventTime = convertTimeunit(eventTime, unit, sim@.xData),
+                         clockTime = obj$._clockTime,
+                         ._clockTime = NULL)]
           }
         }
       }
@@ -2533,7 +2553,7 @@ setReplaceMethod(
 #' @return A \code{simList} object.
 #'
 #' @include simList-class.R
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @keywords internal
 #' @rdname addDepends
 #'
@@ -2584,7 +2604,7 @@ setMethod(
 #'
 #' @export
 #' @include simList-class.R
-#' @family functions to access elements of a \code{simList} object
+#' @family functions to access elements of a 'simList' object
 #' @rdname packages
 #'
 #' @author Alex Chubaty & Eliot McIntire
@@ -2665,7 +2685,6 @@ setMethod(
     return(pkgs)
 })
 
-
 ################################################################################
 #' Metadata accessors
 #'
@@ -2700,7 +2719,7 @@ setMethod("inputObjects",
               sim@depends@dependencies[[module]]@inputObjects
             }
             return(out)
-          })
+})
 
 ################################################################################
 #' @inheritParams P
@@ -2731,8 +2750,41 @@ setMethod("outputObjects",
               sim@depends@dependencies[[module]]@outputObjects
             }
             return(out)
-          })
+})
 
+################################################################################
+#' @inheritParams P
+#' @include simList-class.R
+#' @export
+#' @rdname simList-accessors-metadata
+#' @aliases simList-accessors-metadata
+#'
+setGeneric("outputObjectNames", function(sim, module) {
+  standardGeneric("outputObjectNames")
+})
+
+#' @export
+#' @rdname simList-accessors-metadata
+#' @aliases simList-accessors-metadata
+setMethod("outputObjectNames",
+          signature = "simList",
+          definition = function(sim, module) {
+            out <- if (NROW(modules(sim)) > 0) {
+              outObjs <- outputObjects(sim)
+              allObjNames <- if (is(outObjs, "list")) {
+                allObjNames <- lapply(outObjs, function(x) x$objectName)
+                if (!missing(module)) {
+                  allObjNames <- allObjNames[[module]]
+                }
+                allObjNames
+              } else {
+                list(outObjs$objectName)
+              }
+            } else {
+              NULL
+            }
+            return(out)
+})
 
 ################################################################################
 #' @inheritParams P
@@ -2763,8 +2815,7 @@ setMethod("reqdPkgs",
               unlist(sim@depends@dependencies[[module]]@reqdPkgs)
             }
             return(out)
-          })
-
+})
 
 ################################################################################
 #' @inheritParams P
@@ -2795,8 +2846,7 @@ setMethod("documentation",
               sim@depends@dependencies[[module]]@documentation
             }
             return(out)
-          })
-
+})
 
 ################################################################################
 #' @param package For compatibility with \code{\link[utils]{citation}}. This can be
@@ -2830,7 +2880,7 @@ setMethod("citation",
               package@depends@dependencies[[module]]@citation
             }
             return(out)
-          })
+})
 
 #' @export
 #' @rdname simList-accessors-metadata
@@ -2839,9 +2889,7 @@ setMethod("citation",
           signature = "character",
           definition = function(package, lib.loc, auto, module) {
             utils::citation(package = package, lib.loc = lib.loc, auto = auto)
-          })
-
-
+})
 
 ################################################################################
 #' @inheritParams times
@@ -2869,7 +2917,6 @@ elapsedTime.simList <- function(x, byEvent = TRUE, ...) {
     ret <- comp[, list(elapsedTime = sum(diffTime)), by = theBy]
   } else {
     ret <- NULL
-
   }
   return(ret)
 }