+++ tests/testthat/test-load.R
@@ -411,3 +411,31 @@ test_that("test-load.R: interval loading of objects from .GlobalEnv", {
   expect_identical(a2, a$a)
 
 })
+
+
+test_that("Filenames for simList", {
+  testInitOut <- testInit(c("raster"), tmpFileExt = c(".tif", ".grd", ".tif", ".tif", ".grd"),
+                          opts = list("reproducible.ask" = FALSE))
+
+  on.exit({
+    testOnExit(testInitOut)
+    options(opts)
+    rm(s)
+  }, add = TRUE)
+
+  s <- simInit()
+  s$r <- raster(extent(0,10,0,10), vals = 1, res = 1)
+  s$r2 <- raster(extent(0,10,0,10), vals = 1, res = 1)
+  s$r <- writeRaster(s$r, filename = tmpfile[1], overwrite = TRUE)
+  s$r2 <- writeRaster(s$r2, filename = tmpfile[3], overwrite = TRUE)
+  s$s <- stack(s$r, s$r2)
+  s$b <- writeRaster(s$s, filename = tmpfile[5], overwrite = TRUE)
+
+  Fns <- Filenames(s)
+
+  expect_true(identical(Fns$b, filename(s$b)))
+  expect_true(identical(Fns$r, filename(s$r)))
+  expect_true(identical(Fns$r2, filename(s$r2)))
+  expect_true(identical(Fns$s, sapply(seq_len(nlayers(s$s)), function(rInd) filename(s$s[[rInd]]))))
+
+})